Correct type nEff to int

+++ src/main/java/org/gridsuite/study/server/dto/dynamicsimulation/dynawaltz/solver/SimSolverInfos.java
@@ -32,7 +32,7 @@ public class SimSolverInfos extends AbstractSolverInfos {
     private double kReduceStep;
 
     @JsonProperty("nEff")
-    private double nEff;
+    private int nEff;
 
     @JsonProperty("nDeadband")
     private int nDeadband;
